Support `bin` set to string
Where a package has been `npm link`-ed into a dependency tree rather than installed from npm then it's possible that the bin property is set to a strin, which would resolve to a binary with the same name as the package. In this case, allow this to be used.