Handle modules which are not in the parent tree of `witch`
If calling `witch` from a module where witch is installed as a sibling in the same node_modules directory, but the module being searched for was installed in a child directory of the callsite, then witch would fail because the module under search was not present in its `paths`.

Instead do a manual findup from the callsite to get the calling file's full dependency tree.